# Override and short-circuit Chunk.flatten