@diesalbla 2022-09-22
There are other methods that could be optimised for `Chunk`, as well.